requirements.txt
python-dotenv
pyyaml
requests
tqdm
pandas
beautifulsoup4
feedparser
snscrape
pytrends
python-pptx
jinja2
orjson
openai>=1.40.0
python-docx